# swagger.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  description: API for user authentication and product management
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /product/{id}:
    get:
      summary: Get product details by ID
      description: This endpoint retrieves the details of a product by its ID.
      tags:
        - Product Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the product to retrieve.
      responses:
        "200":
          description: The details of the requested product.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 101
                  name:
                    type: string
                    example: "Product A"
                  description:
                    type: string
                    example: "High-quality product"
                  price:
                    type: number
                    example: 199.99
                  stock:
                    type: integer
                    example: 10
                  categoryId:
                    type: integer
                    example: 1
                  imageUrl:
                    type: string
                    example: "https://example.com/images/product_a.png"
        "422":
          description: Error fetching products
...
  /product/{:id}:
    delete:
      summary: Delete a product by ID
      description: This endpoint deletes a product by its ID.
      tags:
        - Product Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: :id
          required: true
          schema:
            type: integer
          description: The ID of the product to be deleted.
      responses:
        "200":
          description: Confirmation message that the product was deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted successfully"
        "404":
          description: Product not found
        "422":
          description: Error deleting product
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT